timer, feat: support beyond 2^31 delay (#357)

+++ fibjs/include/ifs/global.h
@@ -39,8 +39,8 @@ public:
     static result_t clearInterval(Timer_base* t);
     static result_t clearTimeout(Timer_base* t);
     static result_t clearImmediate(Timer_base* t);
-    static result_t setInterval(v8::Local<v8::Function> callback, int32_t timeout, obj_ptr<Timer_base>& retVal);
-    static result_t setTimeout(v8::Local<v8::Function> callback, int32_t timeout, obj_ptr<Timer_base>& retVal);
+    static result_t setInterval(v8::Local<v8::Function> callback, double timeout, obj_ptr<Timer_base>& retVal);
+    static result_t setTimeout(v8::Local<v8::Function> callback, double timeout, obj_ptr<Timer_base>& retVal);
     static result_t setImmediate(v8::Local<v8::Function> callback, obj_ptr<Timer_base>& retVal);
     static result_t GC();
     static result_t repl(v8::Local<v8::Array> cmds);
@@ -257,7 +257,7 @@ inline void global_base::s_setInterval(const v8::FunctionCallbackInfo<v8::Value>
     METHOD_OVER(2, 2);
 
     ARG(v8::Local<v8::Function>, 0);
-    ARG(int32_t, 1);
+    ARG(double, 1);
 
     hr = setInterval(v0, v1, vr);
 
@@ -273,7 +273,7 @@ inline void global_base::s_setTimeout(const v8::FunctionCallbackInfo<v8::Value>&
     METHOD_OVER(2, 2);
 
     ARG(v8::Local<v8::Function>, 0);
-    ARG(int32_t, 1);
+    ARG(double, 1);
 
     hr = setTimeout(v0, v1, vr);
